IMF DOTS data loading per a country with all other counterpart countres

+++ fetch_data/precomputed_NEER.py
@@ -3,7 +3,6 @@ import os
 import pandas as pd
 from fredapi import Fred # for FRED BIS
 import sdmx # IMF
-import datetime
 
 from utils.config import *
 
@@ -17,7 +16,7 @@ class PrecomputedNeer:
     def onlyPrecomputed(self):
         pass
 
-    def getNeerSeries(self, ccies):
+    def getNeerSeries(self, ccies, loadFileIfExists):
         return None
 
 class PrecomputeNeerBis(PrecomputedNeer):
@@ -33,23 +32,26 @@ class PrecomputeNeerBis(PrecomputedNeer):
             onlyPrecomputed_bool = all([not(value["FRED"] is None) for value in ccies.values() ])  
             return onlyPrecomputed_bool
 
-    def getNeerSeries(self, ccies):
-        fred_bis_api_key = os.getenv("FRED_API_KEY")
-        fred = Fred(api_key=fred_bis_api_key)
-        NEER_dfs = []
-        for ccy, value in ccies.items():
-            fred_key = value["BIS"]
-            isInclude = value["Include"]
-            NEER_country = None
-            if isInclude:
-                if fred_key is not None:
-                    NEER_country = fred.get_series("NB" + fred_key + "BIS").to_frame()
-                    print(f'fetched currency {ccy} from web')
-                    NEER_country.index.name = "Date"
-                    NEER_country.columns = [ccy]
-                    NEER_dfs.append(NEER_country)
-        NEER_df = pd.concat(NEER_dfs, axis=1)
-        NEER_df.to_excel(self.storeFilename, index=True)
+    def getNeerSeries(self, ccies, loadFileIfExists):
+        if loadFileIfExists and os.path.exists(self.storeFilename):
+            NEER_df = pd.read_excel(self.storeFilename, index_col = "Date")
+        else:
+            fred_bis_api_key = os.getenv("FRED_API_KEY")
+            fred = Fred(api_key=fred_bis_api_key)
+            NEER_dfs = []
+            for ccy, value in ccies.items():
+                fred_key = value["BIS"]
+                isInclude = value["Include"]
+                NEER_country = None
+                if isInclude:
+                    if fred_key is not None:
+                        NEER_country = fred.get_series("NB" + fred_key + "BIS").to_frame()
+                        print(f'fetched currency {ccy} from web')
+                        NEER_country.index.name = "Date"
+                        NEER_country.columns = [ccy]
+                        NEER_dfs.append(NEER_country)
+            NEER_df = pd.concat(NEER_dfs, axis=1)
+            NEER_df.to_excel(self.storeFilename, index=True)
         return NEER_df
     
 class PrecomputeNeerImf(PrecomputedNeer):
@@ -66,29 +68,32 @@ class PrecomputeNeerImf(PrecomputedNeer):
             onlyPrecomputed_bool = all([not(value["IMF"] is None) for value in ccies.values() ])  
             return onlyPrecomputed_bool
 
-    def getNeerSeries(self, ccies):
-        IMF_DATA = sdmx.Client('IMF_DATA')
-        res_NEER = "EER"
+    def getNeerSeries(self, ccies, loadFileIfExists):
+        if loadFileIfExists and os.path.exists(self.storeFilename):
+            NEER_df = pd.read_excel(self.storeFilename, index_col = "Date")
+        else:
+            IMF_DATA = sdmx.Client('IMF_DATA')
+            res_NEER = "EER"
 
-        NEER_dfs = []
-        for ccy, value in ccies.items():
-            imf_key = value["IMF"]
-            key_NEER = f"{imf_key}.NEER_IX_RY2010_ACW.{self.freq}"
-            isInclude = value["Include"]
-            NEER_country = None
-            if isInclude:
-                if imf_key is not None:
-                    data_msg = IMF_DATA.data(res_NEER, key=key_NEER, params={"startPeriod": self.startYear})
-                    print(f'fetched currency {ccy} from web')
-                    NEER_country = sdmx.to_pandas(data_msg)
-                    NEER_country_flat_index = NEER_country.reset_index()
-                    NEER_country_flat_index["Date"] = NEER_country_flat_index['TIME_PERIOD'].apply(lambda x: 
-                        pd.to_datetime(x + "-01", format="%Y-M%m-%d")
-                    )
-                    NEER_country_new_index = NEER_country_flat_index.set_index("Date")[["value"]]
-                    NEER_country_new_index.columns = [ccy]
-                    NEER_dfs.append(NEER_country_new_index)
-        NEER_df = pd.concat(NEER_dfs, axis=1)
-        NEER_df.to_excel(self.storeFilename, index=True)
+            NEER_dfs = []
+            for ccy, value in ccies.items():
+                imf_key = value["IMF"]
+                key_NEER = f"{imf_key}.NEER_IX_RY2010_ACW.{self.freq}"
+                isInclude = value["Include"]
+                NEER_country = None
+                if isInclude:
+                    if imf_key is not None:
+                        data_msg = IMF_DATA.data(res_NEER, key=key_NEER, params={"startPeriod": self.startYear})
+                        print(f'fetched currency {ccy} from web')
+                        NEER_country = sdmx.to_pandas(data_msg)
+                        NEER_country_flat_index = NEER_country.reset_index()
+                        NEER_country_flat_index["Date"] = NEER_country_flat_index['TIME_PERIOD'].apply(lambda x: 
+                            pd.to_datetime(x + "-01", format="%Y-M%m-%d")
+                        )
+                        NEER_country_new_index = NEER_country_flat_index.set_index("Date")[["value"]]
+                        NEER_country_new_index.columns = [ccy]
+                        NEER_dfs.append(NEER_country_new_index)
+            NEER_df = pd.concat(NEER_dfs, axis=1)
+            NEER_df.to_excel(self.storeFilename, index=True)
         return NEER_df
